@@ -1,5 +1,5 @@
 module github.com/example/golang-app
 
-go 1.23.4
+go 1.24.0
 
 require github.com/gorilla/mux v1.8.0
