@@ -1,3 +1,3 @@
 module github.com/example/golang-app
 
-go 1.23.4
+go 1.24.0
