@@ -1,155 +1,118 @@
 package gcs
 
 import (
-	"context"
-	"io"
-	"os"
-	"path/filepath"
-
 	"cloud.google.com/go/storage"
+	"context"
+	"fmt"
 	"github.com/SAP/jenkins-library/pkg/log"
 	"github.com/pkg/errors"
 	"google.golang.org/api/iterator"
 	"google.golang.org/api/option"
+	"io"
+	"os"
+	"path/filepath"
 )
 
 // Client is an interface to mock gcsClient
 type Client interface {
-	UploadFile(bucketID string, sourcePath string, targetPath string) error
-	DownloadFile(bucketID string, sourcePath string, targetPath string) error
-	ListFiles(bucketID string) ([]string, error)
+	UploadFile(ctx context.Context, bucketID string, sourcePath string, targetPath string) error
+	DownloadFile(ctx context.Context, bucketID string, sourcePath string, targetPath string) error
+	ListFiles(ctx context.Context, bucketID string) ([]string, error)
 	Close() error
 }
 
-// gcsClient provides functions to interact with google cloud storage API
+// gcsClient provides functions to interact with Google Cloud Storage API
 type gcsClient struct {
-	context       context.Context
-	envVars       []EnvVar
-	client        storage.Client
-	clientOptions []option.ClientOption
-	openFile      func(name string) (io.ReadCloser, error)
-	createFile    func(name string) (io.WriteCloser, error)
+	gcs        storage.Client
+	gcsOptions []option.ClientOption
+	openFile   func(name string) (io.ReadCloser, error)
+	createFile func(name string) (io.WriteCloser, error)
 }
 
-// EnvVar defines an  environment variable incl. information about a potential modification to the variable
-type EnvVar struct {
-	Name     string
-	Value    string
-	Modified bool
-}
+type clientOptions func(*gcsClient)
 
-type gcsOption func(*gcsClient)
-
-// WithEnvVars initializes env variables in gcsClient
-func WithEnvVars(envVars []EnvVar) gcsOption {
-	return func(g *gcsClient) {
-		g.envVars = envVars
-	}
-}
-
-// WithEnvVars initializes the openFile function in gcsClient
-func WithOpenFileFunction(openFile func(name string) (io.ReadCloser, error)) gcsOption {
-	return func(g *gcsClient) {
-		g.openFile = openFile
-	}
-}
-
-// WithEnvVars initializes the createFile function in gcsClient
-func WithCreateFileFunction(createFile func(name string) (io.WriteCloser, error)) gcsOption {
-	return func(g *gcsClient) {
-		g.createFile = createFile
-	}
-}
-
-// WithEnvVars initializes the Google Cloud Storage client options
-func WithClientOptions(opts ...option.ClientOption) gcsOption {
-	return func(g *gcsClient) {
-		g.clientOptions = append(g.clientOptions, opts...)
-	}
-}
-
-// Init intitializes the google cloud storage client
-func NewClient(opts ...gcsOption) (*gcsClient, error) {
-	var (
-		defaultOpenFile   = openFileFromFS
-		defaultCreateFile = createFileOnFS
-	)
-
-	ctx := context.Background()
-	gcsClient := &gcsClient{
-		context:    ctx,
-		openFile:   defaultOpenFile,
-		createFile: defaultCreateFile,
+// NewClient initializes the Google Cloud Storage client with the provided options
+func NewClient(keyFile, token string, opts ...clientOptions) (Client, error) {
+	client := &gcsClient{
+		openFile:   openFileFromFS,
+		createFile: createFileOnFS,
 	}
 
-	// options handling
+	// Apply options
 	for _, opt := range opts {
-		opt(gcsClient)
+		opt(client)
 	}
 
-	gcsClient.prepareEnv()
-	client, err := storage.NewClient(ctx, gcsClient.clientOptions...)
+	gcs, err := initGcsClient(context.Background(), keyFile, token, client.gcsOptions...)
 	if err != nil {
-		return nil, errors.Wrapf(err, "bucket connection failed: %v", err)
+		return nil, err
 	}
-	gcsClient.client = *client
-	return gcsClient, nil
+
+	client.gcs = *gcs
+	return client, nil
 }
 
-// UploadFile uploads a file into a google cloud storage bucket
-func (g *gcsClient) UploadFile(bucketID string, sourcePath string, targetPath string) error {
-	target := g.client.Bucket(bucketID).Object(targetPath).NewWriter(g.context)
-	log.Entry().Debugf("uploading %v to %v", sourcePath, targetPath)
-	sourceFile, err := g.openFile(sourcePath)
+func (cl *gcsClient) UploadFile(ctx context.Context, bucketID string, sourcePath string, targetPath string) error {
+	sourcePath = filepath.Clean(sourcePath)
+	log.Entry().Debugf("Uploading %v to %v\n", sourcePath, targetPath)
+
+	sourceFile, err := cl.openFile(sourcePath)
 	if err != nil {
-		return errors.Wrapf(err, "could not open source file: %v", err)
+		return fmt.Errorf("failed to open source file: %w", err)
 	}
 	defer sourceFile.Close()
 
-	if err := g.copy(sourceFile, target); err != nil {
-		return errors.Wrapf(err, "upload failed: %v", err)
+	target := cl.gcs.Bucket(bucketID).Object(targetPath).NewWriter(ctx)
+	task := func(ctx context.Context) error {
+		if _, err = io.Copy(target, sourceFile); err != nil {
+			return fmt.Errorf("upload failed: %w", err)
+		}
+		// Google API errors (like 401, 403 and etc) are returned by Close() method, for some reason.
+		return target.Close()
 	}
 
-	if err := target.Close(); err != nil {
-		return errors.Wrapf(err, "closing bucket failed: %v", err)
-	}
-	return nil
+	return retryWithLogging(ctx, log.Entry(), task, initialBackoff, maxRetries, retryMultiplier)
 }
 
-// DownloadFile downloads a file from a google cloud storage bucket
-func (g *gcsClient) DownloadFile(bucketID string, sourcePath string, targetPath string) error {
-	log.Entry().Debugf("downloading %v to %v\n", sourcePath, targetPath)
-	gcsReader, err := g.client.Bucket(bucketID).Object(sourcePath).NewReader(g.context)
-	if err != nil {
-		return errors.Wrapf(err, "could not open source file from a google cloud storage bucket: %v", err)
-	}
+// DownloadFile downloads a file from a Google Cloud Storage bucket
+func (cl *gcsClient) DownloadFile(ctx context.Context, bucketID, sourcePath, targetPath string) error {
+	targetPath = filepath.Clean(targetPath)
+	log.Entry().Debugf("Downloading %v to %v\n", sourcePath, targetPath)
 
-	targetWriter, err := g.createFile(targetPath)
+	target, err := cl.createFile(targetPath)
 	if err != nil {
 		return errors.Wrapf(err, "could not create target file: %v", err)
 	}
-	defer targetWriter.Close()
+	defer target.Close()
 
-	if err := g.copy(gcsReader, targetWriter); err != nil {
-		return errors.Wrapf(err, "download failed: %v", err)
+	source, err := cl.gcs.Bucket(bucketID).Object(sourcePath).NewReader(ctx)
+	if err != nil {
+		return fmt.Errorf("failed to open source file: %w", err)
 	}
-	if err := gcsReader.Close(); err != nil {
-		return errors.Wrapf(err, "closing bucket failed: %v", err)
+
+	task := func(ctx context.Context) error {
+		if _, err = io.Copy(target, source); err != nil {
+			return fmt.Errorf("download failed: %w", err)
+		}
+
+		// Google API errors (like 401, 403 and etc) are returned by Close() method, for some reason.
+		return source.Close()
 	}
-	return nil
+
+	return retryWithLogging(ctx, log.Entry(), task, initialBackoff, maxRetries, retryMultiplier)
 }
 
-// ListFiles lists all files in certain GCS bucket
-func (g *gcsClient) ListFiles(bucketID string) ([]string, error) {
+// ListFiles lists all files in a specified Google Cloud Storage bucket
+func (cl *gcsClient) ListFiles(ctx context.Context, bucketID string) ([]string, error) {
 	fileNames := []string{}
-	it := g.client.Bucket(bucketID).Objects(g.context, nil)
+	it := cl.gcs.Bucket(bucketID).Objects(ctx, nil)
 	for {
 		attrs, err := it.Next()
-		if err == iterator.Done {
+		if errors.Is(err, iterator.Done) {
 			break
 		}
 		if err != nil {
-			return nil, err
+			return nil, errors.Wrap(err, "could not list files")
 		}
 		fileNames = append(fileNames, attrs.Name)
 	}
@@ -157,64 +120,71 @@ func (g *gcsClient) ListFiles(bucketID string) ([]string, error) {
 }
 
 // Close closes the client and removes previously set environment variables
-func (g *gcsClient) Close() error {
-	if err := g.client.Close(); err != nil {
+func (cl *gcsClient) Close() error {
+	if err := cl.gcs.Close(); err != nil {
 		return err
 	}
-	if err := g.cleanupEnv(); err != nil {
-		return err
-	}
+
 	return nil
 }
 
-func (g *gcsClient) copy(source io.Reader, target io.Writer) error {
-	if _, err := io.Copy(target, source); err != nil {
-		return err
-	}
-	return nil
-}
+// TODO: consider refactoring to avoid keeping functions that are used only for integration testing in the production code
+// Below functions WithOpenFileFunction, WithCreateFileFunction and WithGCSClientOptions are
+// used only for unit/integration testing.
 
-// prepareEnv sets required environment variables in case they are not set yet
-func (g *gcsClient) prepareEnv() {
-	for key, env := range g.envVars {
-		g.envVars[key].Modified = setenvIfEmpty(env.Name, env.Value)
+// WithOpenFileFunction sets the openFile function in gcsClient
+func WithOpenFileFunction(openFile func(name string) (io.ReadCloser, error)) clientOptions {
+	return func(g *gcsClient) {
+		g.openFile = openFile
 	}
 }
 
-// cleanupEnv removes environment variables set by prepareEnv
-func (g *gcsClient) cleanupEnv() error {
-	for _, env := range g.envVars {
-		if err := removeEnvIfPreviouslySet(env.Name, env.Modified); err != nil {
-			return err
-		}
+// WithCreateFileFunction sets the createFile function in gcsClient
+func WithCreateFileFunction(createFile func(name string) (io.WriteCloser, error)) clientOptions {
+	return func(g *gcsClient) {
+		g.createFile = createFile
 	}
-	return nil
 }
 
-func setenvIfEmpty(env, val string) bool {
-	if len(os.Getenv(env)) == 0 {
-		os.Setenv(env, val)
-		return true
+// WithGCSClientOptions sets the Google Cloud Storage client options
+func WithGCSClientOptions(opts ...option.ClientOption) clientOptions {
+	return func(g *gcsClient) {
+		g.gcsOptions = append(g.gcsOptions, opts...)
 	}
-	return false
 }
 
-func removeEnvIfPreviouslySet(env string, previouslySet bool) error {
-	if previouslySet {
-		if err := os.Setenv(env, ""); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
+// openFileFromFS and createFileOnFS functions are existing just because of the integration tests
+// TODO: find a better way to mock filesystem operations in the integration tests
+// openFileFromFS opens a file from the filesystem
 func openFileFromFS(name string) (io.ReadCloser, error) {
 	return os.Open(name)
 }
 
+// createFileOnFS creates a file on the filesystem
 func createFileOnFS(name string) (io.WriteCloser, error) {
 	if err := os.MkdirAll(filepath.Dir(name), os.ModePerm); err != nil {
 		return nil, err
 	}
 	return os.Create(name)
+}
+
+// NewClientLegacy is also still here because of integration tests
+func NewClientLegacy(opts ...clientOptions) (Client, error) {
+	client := &gcsClient{
+		openFile:   openFileFromFS,
+		createFile: createFileOnFS,
+	}
+
+	// Apply options
+	for _, opt := range opts {
+		opt(client)
+	}
+
+	gcs, err := storage.NewClient(context.Background(), client.gcsOptions...)
+	if err != nil {
+		return nil, err
+	}
+
+	client.gcs = *gcs
+	return client, nil
 }
